+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Changed `Request.Params` is be a getter method.
 - Added `Request.WithParams` to set params.
+- Replaced `treemux.WithPathSource(URLPath)` with `treemux.UseURLPath()`.
 
 ## v0.7.0 - Jun 28 2021